requirements.txt
# UI
PyQt6>=6.5.0

# Audio processing
pydub>=0.25.1
librosa>=0.10.0

# Text processing
markitdown>=0.1.0

# Speech-to-Text
openai-whisper>=20231117

# Text-to-Speech
kokoro-onnx>=0.1.0
soundfile>=0.12.1
sounddevice>=0.4.6

# Utilities
numpy>=1.24.0
